+++ forui/CHANGELOG.md
@@ -19,6 +19,9 @@
 * **Breaking:** Remove `onChange` parameter from `FSelectTile`. This was an artifact from early prototyping that was 
   unfortunately overlooked and did not remove.
 
+### Fixes
+
+* Resolved an issue where `FTappable` remained in a hovered or touched state when its `onPress`/`onLongPress` callbacks were nulled after being non-null.
 
 ## 0.8.0